minor bug fixes.
variables can now be assigned the values of other variables.
When evaluating a infix queue that is only a single variable, it now properly returns the value of the variable instead of the variable itself.
printing name is now properly formatted, with a new line at the end.
now exiting when variable is declared but not initiated (invalid input, should never happen).